# Conflicts:
#	Web/src/main/java/com/engjoy/config/SecurityConfig.java
#	Web/src/main/java/com/engjoy/control/HomeController.java
#	Web/src/main/java/com/engjoy/dto/PasswordResetDto.java
#	Web/src/main/java/com/engjoy/repository/AccountRepository.java
#	Web/src/main/java/com/engjoy/repository/AchievementDescRepository.java
#	Web/src/main/java/com/engjoy/repository/UserAchievementRepository.java
#	Web/src/main/java/com/engjoy/repository/UserGameDataRepository.java
#	Web/src/main/java/com/engjoy/service/AccountService.java
#	Web/src/main/resources/templates/login.html

+++ Web/src/main/java/com/engjoy/repository/UserAchievementRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface UserAchievementRepository extends JpaRepository<UserAchievement, Integer> {
+public interface UserAchievementRepository extends JpaRepository<UserAchievement, Long> {
 }
